REFACT - Improvements to signature update - added `char_handler` class to gather duplicated character escape code; - removed duplicated character constants and escape functions from `fido/fido.py` and `fido/prepare.py`; - refactored sig loading code in `fido/fido.py` to fail gracefully when sig file parsing fails; and - update `pronomContainerSignature` value in `fido/conf/versions.xml` to reflect new container sig file.
