+++ fido/conf/formats-v95.xml
@@ -55162,4 +55162,3 @@ f &lt;space&gt; 3+ integer values, can be negative, separated by space. Faces co
     </details>
   </format>
 </formats>
-
