+++ fido/conf/versions.xml
@@ -2,7 +2,7 @@
 <versions>
 	<pronomVersion>95</pronomVersion>
 	<pronomSignature>formats-v95.xml</pronomSignature>
-	<pronomContainerSignature>container-signature-20160121.xml</pronomContainerSignature>
+	<pronomContainerSignature>container-signature-20180920.xml</pronomContainerSignature>
 	<fidoExtensionSignature>format_extensions.xml</fidoExtensionSignature>
-	<updateScript>1.3.13</updateScript>
+	<updateScript>beta-1.4</updateScript>
 </versions>
